Remove references to the git repo.
I don't think it's that important to cite the version of the
documentation, and if someone really wants to know (and would
know what git was) they could find out from the url.

If we do want to say it's from git, as well as "Pre-release"
we should do it once for the section, not per line.


git-svn-id: http://svn.xiph.org/websites/opus-codec.org@18387 0101bb08-14d6-0310-b084-bc0e0c8e3800