Mega-commit: Convert website to use Jekyll static site generator
This converts the website to use the Jekyll static site generator
which offers more flexibility and the benefit or real static html
pages, while content can be written in markdown or html, depending
on what you like the most.
Additionally this uses a new design for the website, which looks
better and is optimized to look good on mobile devices as well.